[All] Use a single namespace for all the dictionary-plugin domain objects
There was some redundancy between the namespaces and the dictionary root
keys, and a confusing use of multiple namespaces for the same plugin.
Multiple namespaces were used for selecting the dictionary to use in a
telemetry domain object provide request, while namespace is meant for
distinguishing domain objects from different Open MCT plugins. Instead
of using the namespace for selecting the dictionary, it makes more sense
to use a prefix from the domain object ID. That prefix could be the
dictionary ID:
- Remove previous namespaces and define a single one for all the objects
  in the dictionary-plugin: YARPOPENMCT_DICTIONARY_PLUGIN_NAMESPACE.
- Add dictionary prefix to the telemetry IDs (`icubtelemetry`,
  `processLogging`, etc).
- Use that prefix as a sub-namespace for selecting the dictionary holding
  the domain object properties.
- Use sub-namespace VECTORCOLLECTIONS_DOMAIN_OBJECTS_SUBNAMESPACE defined
  from the vector collections dictionary.

Other minor refactoring:
- Namespaces are defined a single time in `namespacesNtypes.js` file.
- Factorize `openmct.objects.addRoot` operation.